Refined class utils, added some new ones

+++ SpecterLib/SyscallException.hpp
@@ -7,7 +7,7 @@ class SyscallException : public std::system_error
 {
 public:
 	SyscallException(const std::string& error_msg, DWORD error_code = GetLastError()) :
-		std::system_error(error_code, std::system_category(), error_msg + ", error")
+		std::system_error(static_cast<int>(error_code), std::system_category(), error_msg + ", error")
 	{
 	}
 };
